Extract shared utilities, deduplicate session ID extraction, tighten API
Code cleanup based on full codebase review:

1. Extract EmulatorInfrastructure (Core/Hosting):
   - LoadDevCert, GetFreePort, EnsureDevCertTrusted shared between
     Host and TestHost (was duplicated in both)

2. Extract ExtractSessionId helper in ManagementLinkEndpoint:
   - Session ID extraction from request body/ApplicationProperties
     was copy-pasted across HandleRenewSessionLock,
     HandleGetSessionState, HandleSetSessionState

3. Tighten API surface:
   - SenderLinkEndpoint.RouteMessage → internal (only used within assembly)
   - Add InternalsVisibleTo for test project
