+++ src/AzureServiceBusEmulator.Core/AzureServiceBusEmulator.Core.csproj
@@ -10,4 +10,7 @@
   <ItemGroup>
     <PackageReference Include="AMQPNetLite" Version="2.5.1" />
   </ItemGroup>
+  <ItemGroup>
+    <InternalsVisibleTo Include="AzureServiceBusEmulator.Tests" />
+  </ItemGroup>
 </Project>
