+++ src/AzureServiceBusEmulator.Core/Hosting/EmulatorInfrastructure.cs
@@ -1,0 +1,66 @@
+using System.Net;
+using System.Net.Sockets;
+using System.Security.Cryptography.X509Certificates;
+
+namespace AzureServiceBusEmulator.Core.Hosting;
+
+/// <summary>
+/// Shared infrastructure utilities used by both the standalone host and the test fixture.
+/// </summary>
+public static class EmulatorInfrastructure
+{
+    /// <summary>
+    /// Finds an available TCP port on the loopback interface.
+    /// </summary>
+    public static int GetFreePort()
+    {
+        var listener = new TcpListener(IPAddress.Loopback, 0);
+        listener.Start();
+        var port = ((IPEndPoint)listener.LocalEndpoint).Port;
+        listener.Stop();
+        return port;
+    }
+
+    /// <summary>
+    /// Loads the ASP.NET Core HTTPS development certificate from the current user's certificate store.
+    /// Throws if no dev cert is found.
+    /// </summary>
+    public static X509Certificate2 LoadDevCert()
+    {
+        using var store = new X509Store(StoreName.My, StoreLocation.CurrentUser);
+        store.Open(OpenFlags.ReadOnly);
+        // OID 1.3.6.1.4.1.311.84.1.1 identifies ASP.NET Core dev certs
+        var certs = store.Certificates.Find(
+            X509FindType.FindByExtension, "1.3.6.1.4.1.311.84.1.1", validOnly: false);
+        if (certs.Count == 0)
+            throw new InvalidOperationException(
+                "ASP.NET HTTPS development certificate not found. " +
+                "Run 'dotnet dev-certs https --trust' to generate and trust the certificate.");
+        return new X509Certificate2(certs[0]);
+    }
+
+    /// <summary>
+    /// On Linux, ensures the ASP.NET dev cert is in SSL_CERT_DIR so the Azure SDK's
+    /// AMQP TLS client trusts it. dotnet dev-certs --trust places the cert in
+    /// ~/.aspnet/dev-certs/trust but doesn't update SSL_CERT_DIR automatically.
+    /// </summary>
+    public static void EnsureDevCertTrusted()
+    {
+        var trustDir = Path.Combine(
+            Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
+            ".aspnet", "dev-certs", "trust");
+
+        if (!Directory.Exists(trustDir))
+            return;
+
+        var current = Environment.GetEnvironmentVariable("SSL_CERT_DIR") ?? "";
+        if (!current.Contains(trustDir))
+        {
+            var systemCerts = "/usr/lib/ssl/certs";
+            var newValue = string.IsNullOrEmpty(current)
+                ? $"{trustDir}:{systemCerts}"
+                : $"{trustDir}:{current}";
+            Environment.SetEnvironmentVariable("SSL_CERT_DIR", newValue);
+        }
+    }
+}
